feat: add rm function

+++ src/fs/add.js
@@ -1,16 +1,22 @@
-import fs from 'fs';
+import fs from "fs";
 import { log } from "../utils/colorConsole/colorConsole.js";
-import path from 'path';
+import path from "path";
 
 export const add = (newFileName, currentDir) => {
   log.cyan("add запускаем add");
   const filePath = path.join(currentDir, newFileName);
   try {
-    fs.writeFile(filePath, '', (err) => {
-      if (err) throw err;
-      console.log(`${newFileName} был успешно создан в каталоге  ${currentDir}!`);
-    }); 
+    if (fs.existsSync(filePath)) {
+      log.red(`${newFileName} уже существует в каталоге!`);
+    } else {
+      fs.writeFile(filePath, "", (err) => {
+        if (err) throw err;
+        console.log(
+          `${newFileName} был успешно создан в каталоге  ${currentDir}!`
+        );
+      });
+    }
   } catch (err) {
-    log.red(`Ошибка при создании файла: ${ err }`);
+    log.red(`Ошибка при создании файла: ${err}`);
   }
-}
+};
